+++ myokit/_unit.py
@@ -544,7 +544,11 @@ class Unit(object):
         Existing entries are overwritten without warning.
         """
         # Overwrite existing entries without warning
-        Unit._preferred_representations[unit] = rep
+        if not isinstance(unit, myokit.Unit):
+            raise ValueError(
+                'Second argument to register_preferred_representation must be'
+                ' a myokit.Unit')
+        Unit._preferred_representations[unit] = str(rep)
 
     def __repr__(self):
         """
@@ -560,8 +564,8 @@ class Unit(object):
         return Unit(list(self._x), self._m + math.log10(other))
 
     def __rtruediv__(self, other):
-        # Evaluates ``other / self``, where other is not a unit when future
-        # division is active.
+        # Evaluates ``other / self``, where other is not a unit, in Python 3
+        # or when future division is active.
 
         return Unit([-a for a in self._x], math.log10(other) - self._m)
 
@@ -825,7 +829,7 @@ class Quantity(object):
         return self._value
 
     def __hash__(self):
-        return self._str
+        return hash(self._value) + hash(self._unit)
 
     def __mul__(self, other):
         if not isinstance(other, Quantity):